Move `CalculateTrustedDeviceRequirement` and tests to a separate file
